MIPS64: Add platform-specific linux loader
Unlike other UEFI platforms, Loongson's MIPS64 EFI implementation
does not load the kernel as a PE executable. We need to load it as
an ELF image, and therefore cannot use the common loading logic.

Running the relocator in an EFI environment means that we need to
ask the firmware to map memory before the relocator allocates memory.
Here we do it using grub_memalign, and free the region before calling
the relocator.

Based on code from Heiher <r@hev.cc> found at:

https://github.com/yetist/grub

Signed-off-by: Henry Chen <henry.chen@oss.cipunited.com>